Postgresql, DB workers expanded, AVG tables added
The database was changed from SQLite3 to postgres due to postgres's
ability of handling concurrent reads and writes, resulting in a
significant performance boost. All data was migrated over to the new
psotgresql database.

New tables to accommodate averages for every thirty minutes, two hours,
and six hours, with associated changes in DB nuking and workers for said
tables.

New workers introduced such as a 30 minute averaging worker, 2 hours
averaging worker, a 6 hour averaging worker, a worker used to clean the
DB, and a worker to fill all of the averaging tables.

A new method (database) in the static pages viewer to display a quick
overlook of the average databases.

Aesthetic changes in the homepage such as the bar graph now being
displayed over six hour intervals instead of five minute intervals. PPC
altcurrency was removed in favour of XPM currency for display in the bar
graph.