BTC-E API futher cleaned, API fixes
The BTC-E changes were not properly implimented within hardworker,
resulting in errors when the worker ran. The fix was just changing
the function call names.

A delay of 300 ms was also added in BTC-E fetches due to BTC-E not
seeming to like me using their API every minute even though they
said it was alright.