JS no longer includes IRB, GON is used instead.
The line chart on the homepage now displays the actual date. The issue
earlier was due to rails seeming to send the time in a normal format instead
of milliseconds since 1967 or whatever the date was. The JS now converts the
time to milliseconds properly.

The IRB for the JS homepage not uses the GON gem instead, preparing for the
transfer of the JS to an external JS file, allowing the JS to work within
the rails asset pipeline.

The data gotten from the timedassets table for the line chart is now done
using pluck and last, decreasing the data fetch time down to sub 2 ms. Check
notes for more information.